Import of Blocks at launch of Snap4Arduino
If we want to have an import of blocks at the launch of snap4arduino, we can use it. This method is better than the load one, as the specific path of xml not needed here. Just need to add xml in the same folder, where we have stored tools.xml.